fix_bug_in_disconnect_whatsapp

+++ server/api/v1/companyprofile/company.controller.js
@@ -386,7 +386,10 @@ exports.deleteWhatsAppInfo = function (req, res) {
           //     sandboxCode: req.body.sandboxCode
           //   }}
           // }
-          if (req.body.type === 'Disconnect') {
+          if (req.body.type === 'Disconnect' && !req.body.connected) {
+            updated = {'whatsApp.connected': req.body.connected, 'whatsApp.dateDisconnected': req.body.Date}
+          }
+          else if (req.body.type === 'Disconnect') {
             updated = {$unset: {whatsApp: 1}}
           } else {
             updated = {twilioWhatsApp: {
@@ -404,10 +407,19 @@ exports.deleteWhatsAppInfo = function (req, res) {
             })
         },
         function (callback) {
-          if (req.body.type === 'Disconnect') {
-            utility.callApi(`user/update`, 'post', {query: {_id: req.user._id}, newPayload: {platform: 'messenger'}, options: {}})
+          utility.callApi(`user/update`, 'post', {query: {_id: req.user._id}, newPayload: {platform: 'messenger'}, options: {}})
+            .then(data => {
+              callback(null)
+            })
+            .catch(err => {
+              callback(err)
+            }) 
+        },
+        function (callback) {
+          if (req.body.type === 'Disconnect' && req.body.connected) {
+            utility.callApi(`whatsAppContacts/deleteMany`, 'delete', {companyId: req.user.companyId})
               .then(data => {
-                callback(null)
+                callback(null, data)
               })
               .catch(err => {
                 callback(err)
@@ -417,40 +429,56 @@ exports.deleteWhatsAppInfo = function (req, res) {
           }
         },
         function (callback) {
-          utility.callApi(`whatsAppContacts/deleteMany`, 'delete', {companyId: req.user.companyId})
-            .then(data => {
-              callback(null, data)
-            })
-            .catch(err => {
-              callback(err)
-            })
+          if (req.body.type === 'Disconnect' && req.body.connected) {
+            let query = {
+              purpose: 'deleteMany',
+              match: {companyId: req.user.companyId}
+            }
+            utility.callApi(`whatsAppBroadcasts`, 'delete', query, 'kiboengage')
+              .then(data => {
+                callback(null, data)
+              })
+              .catch(err => {
+                callback(err)
+              })
+          } else {
+            callback(null)
+          }
         },
         function (callback) {
-          let query = {
-            purpose: 'deleteMany',
-            match: {companyId: req.user.companyId}
+          if (req.body.type === 'Disconnect' && req.body.connected) {
+            let query = {
+              purpose: 'deleteMany',
+              match: {companyId: req.user.companyId}
+            }
+            utility.callApi(`whatsAppBroadcastMessages`, 'delete', query, 'kiboengage')
+              .then(data => {
+                callback(null, data)
+              })
+              .catch(err => {
+                callback(err)
+              })
+          } else {
+            callback(null)
           }
-          utility.callApi(`whatsAppBroadcasts`, 'delete', query, 'kiboengage')
-            .then(data => {
-              callback(null, data)
-            })
-            .catch(err => {
-              callback(err)
-            })
         },
         function (callback) {
-          let query = {
-            purpose: 'deleteMany',
-            match: {companyId: req.user.companyId}
+          if (req.body.type === 'Disconnect' && req.body.connected) {
+            let query = {
+              purpose: 'deleteMany',
+              match: {companyId: req.user.companyId}
+            }
+            utility.callApi(`whatsAppChat`, 'delete', query, 'kibochat')
+              .then(data => {
+                data = data[0]
+                callback(null, data)
+              })
+              .catch(err => {
+                callback(err)
+              })
+          } else {
+            callback(null)
           }
-          utility.callApi(`whatsAppChat`, 'delete', query, 'kibochat')
-            .then(data => {
-              data = data[0]
-              callback(null, data)
-            })
-            .catch(err => {
-              callback(err)
-            })
         }
       ], 10, function (err, results) {
         if (err) {
